+++ lib/karafka/routing/features/manual_offset_management/config.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Karafka
+  module Routing
+    module Features
+      class ManualOffsetManagement < Base
+        # Config for manual offset management feature
+        Config = Struct.new(:active, keyword_init: true) do
+          alias_method :active?, :active
+        end
+      end
+    end
+  end
+end