+++ spec/lib/karafka/pro/active_job/dispatcher_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'karafka/pro/active_job/dispatcher'
-
 RSpec.describe_current do
   subject(:dispatcher) { described_class.new }