+++ spec/lib/karafka/pro/processing/partitioner_spec.rb
@@ -1,8 +1,5 @@
 # frozen_string_literal: true
 
-require 'karafka/pro/processing/partitioner'
-require 'karafka/pro/routing/topic_extensions'
-
 RSpec.describe_current do
   it { expect(described_class).to be < Karafka::Processing::Partitioner }
 
@@ -14,7 +11,8 @@ RSpec.describe_current do
   let(:messages) { Array.new(100) { build(:messages_message) } }
 
   before do
-    topic.singleton_class.prepend Karafka::Pro::Routing::TopicExtensions
+    topic.singleton_class.prepend Karafka::Pro::Routing::Features::VirtualPartitions::Topic
+    topic.singleton_class.prepend Karafka::Pro::Routing::Features::LongRunningJob::Topic
     ::Karafka::App.config.concurrency = concurrency
   end